moving FloatingHealthBar to widgets.py
